Move icon pill to beatmap card-specific namespace
